requirements.txt
anthropic>=0.40.0
streamlit>=1.40.0
tavily-python>=0.3.9
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.1.0
resend>=2.0.0
